resize sidebar + cleanup app.tsx
